Refactor eBay integration models
