+++ OneSila/sales_channels/integrations/ebay/models/orders.py
@@ -1,0 +1,11 @@
+from sales_channels.models.orders import RemoteOrder, RemoteOrderItem
+
+
+class EbayOrder(RemoteOrder):
+    """eBay order model."""
+    pass
+
+
+class EbayOrderItem(RemoteOrderItem):
+    """eBay order item model."""
+    pass
